DomainFinder/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.06.15
chardet==5.0.0
colorama==0.4.4
dnspython==2.2.1
exrex==0.10.5
fire==0.4.0
future==0.18.2
idna==3.3
loguru==0.6.0
PySocks==1.7.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2
SQLAlchemy==1.3.22
tenacity==8.0.1
termcolor==1.1.0
tqdm==4.64.0
treelib==1.6.1
urllib3==1.26.9
win32-setctime==1.1.0
dnspython==2.2.1
export
argparse
exrex
async_timeout
selenium
dnslib
threadpool
tldextract